+++ src/main/java/com/raoulvdberge/refinedstorage/gui/GuiCraftingPreview.java
@@ -78,10 +78,12 @@ public class GuiCraftingPreview extends GuiBase {
 
     @Override
     public void update(int x, int y) {
-        scrollbar.setEnabled(getRows() > VISIBLE_ROWS);
-        scrollbar.setMaxOffset(getRows() - VISIBLE_ROWS);
+        if (scrollbar != null) {
+            scrollbar.setEnabled(getRows() > VISIBLE_ROWS);
+            scrollbar.setMaxOffset(getRows() - VISIBLE_ROWS);
+        }
 
-        if (!startButton.enabled && isCtrlKeyDown() && isShiftKeyDown()) {
+        if (startButton != null && !startButton.enabled && isCtrlKeyDown() && isShiftKeyDown()) {
             startButton.enabled = true;
         }
     }