+++ src/main/java/com/raoulvdberge/refinedstorage/gui/GuiReaderWriter.java
@@ -60,8 +60,10 @@ public class GuiReaderWriter extends GuiBase {
 
     @Override
     public void update(int x, int y) {
-        scrollbar.setEnabled(getRows() > VISIBLE_ROWS);
-        scrollbar.setMaxOffset(getRows() - VISIBLE_ROWS);
+        if (scrollbar != null) {
+            scrollbar.setEnabled(getRows() > VISIBLE_ROWS);
+            scrollbar.setMaxOffset(getRows() - VISIBLE_ROWS);
+        }
 
         if (itemSelected >= getChannels().size()) {
             itemSelected = -1;